+++ kite-core/src/main/kotlin/com/tang/kite/wrapper/query/QueryWrapper.kt
@@ -4,8 +4,8 @@ import com.tang.kite.enumeration.SqlType
 import com.tang.kite.function.SFunction
 import com.tang.kite.mapper.BaseMapper
 import com.tang.kite.sql.Column
-import com.tang.kite.sql.SqlNode
 import com.tang.kite.sql.TableReference
+import com.tang.kite.sql.ast.SqlNode
 import com.tang.kite.sql.dialect.SqlDialect
 import com.tang.kite.sql.statement.LogicalStatement
 import com.tang.kite.sql.statement.SqlStatement